Move json header in src/external/json.hpp
This will allow to install rascal together with all required header
