+++ CMakeLists.txt
@@ -145,11 +145,6 @@ include_directories(
   ${CMAKE_SOURCE_DIR}/src
  )
 
-include_directories(SYSTEM
-  ${CMAKE_SOURCE_DIR}/external
-  )
-
-
 # set up bindings
 if(BUILD_BINDINGS)
   find_package(PythonLibsNew 3 MODULE REQUIRED)
